@@ -26,12 +26,18 @@ Gem::Specification.new do |gem|
   gem.test_files            = %w()
   gem.has_rdoc              = false
 
-  gem.required_ruby_version = '>= 1.8.7'
+  gem.required_ruby_version = '>= 1.9.2'
   gem.requirements          = []
   gem.add_runtime_dependency     'sxp'
   gem.add_runtime_dependency     'rdf'  # Required by sxp
   gem.add_development_dependency 'rspec',           '>= 2.12.0'
   gem.add_development_dependency 'yard' ,           '>= 0.8.3'
   gem.add_development_dependency 'rake'
+
+  # Rubinius has it's own dependencies
+  if RUBY_ENGINE == "rbx" && RUBY_VERSION >= "2.1.0"
+    gem.add_runtime_dependency     "rubysl-rational"
+  end
+
   gem.post_install_message  = nil
 end
