@@ -18,27 +18,27 @@ describe EBNF::LL1::Parser do
   describe "ClassMethods" do
     describe "production" do
       it "adds as a start_handler" do
-        ParserTest.start_handlers.keys.should == [:term]
-        ParserTest.start_handlers[:term].should be_a(Proc)
+        expect(ParserTest.start_handlers.keys).to eq [:term]
+        expect(ParserTest.start_handlers[:term]).to be_a(Proc)
       end
       it "adds as a production_handler" do
-        ParserTest.production_handlers.keys.should == [:term]
-        ParserTest.production_handlers[:term].should be_a(Proc)
+        expect(ParserTest.production_handlers.keys).to eq [:term]
+        expect(ParserTest.production_handlers[:term]).to be_a(Proc)
       end
     end
 
     describe "terminal" do
       it "adds as a terminal_handler" do
-        ParserTest.terminal_handlers.keys.should =~ [:escape, :unescape]
-        ParserTest.terminal_handlers[:escape].should be_a(Proc)
-        ParserTest.terminal_handlers[:unescape].should be_a(Proc)
+        expect(ParserTest.terminal_handlers.keys).to include(:escape, :unescape)
+        expect(ParserTest.terminal_handlers[:escape]).to be_a(Proc)
+        expect(ParserTest.terminal_handlers[:unescape]).to be_a(Proc)
       end
 
       it "adds patterns" do
-        ParserTest.patterns.should =~ [
+        expect(ParserTest.patterns).to include(
           EBNF::LL1::Lexer::Terminal.new(:escape, /escape/),
           EBNF::LL1::Lexer::Terminal.new(:unescape, /unescape/, :unescape => true)
-        ]
+        )
       end
     end
   end
@@ -46,13 +46,13 @@ describe EBNF::LL1::Parser do
   describe "#parse" do
     subject {ParserTest.new}
     it "raises error if no branch table defined" do
-      lambda {subject.parse("foo")}.should raise_error(EBNF::LL1::Parser::Error, "Branch table not defined")
+      expect {subject.parse("foo")}.to raise_error(EBNF::LL1::Parser::Error, "Branch table not defined")
     end
 
     it "raises error if starting production not defined" do
-      lambda {
+      expect {
         subject.parse("foo", nil, :branch => {:a => {:b => ["c"]}})
-      }.should raise_error(EBNF::LL1::Parser::Error, "Starting production not defined")
+      }.to raise_error(EBNF::LL1::Parser::Error, "Starting production not defined")
     end
   end
 end
